Add files via upload
In this article, I have discussed introductory and important aspects of sequence analysis along with multiple sequence alignment and phylogenetic analysis in Bioinformatics. Sequence analysis is the most basic task in bioinformatics. In general, it refers to processing sequence data of DNA or proteins to harness the underlying information about the function, structure, or evolution of the biomolecule. Multiple sequence alignment is also one of the most fundamental tasks in bioinformatics. It often leads to fundamental biological insight into sequence-structure-function relationships of nucleotide or protein sequence families.